+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+- Disable country edition for addresses saved at the profile
+
 ## [1.27.1] - 2024-02-15
 
 ### Fixed